Split steps file
