+++ packages/app/e2e-tests/wdio.conf.ts
@@ -208,7 +208,7 @@ export const config: Options.Testrunner = {
   // If you are using Cucumber you need to specify the location of your step definitions.
   cucumberOpts: {
     // <string[]> (file/dir) require files before executing features
-    require: [`${dir}/step-definitions/steps.ts`],
+    require: [`${dir}/step-definitions/**/*.ts`],
     // <boolean> show full backtrace for errors
     backtrace: false,
     // <string[]> ("extension:module") require files with the given EXTENSION after requiring MODULE (repeatable)
